+++ .backend_env.example
@@ -12,8 +12,6 @@ MAX_REQUESTS_NUMBER=200
 #Private LLM Variables
 PRIVATE=False
 MODEL_PATH=./local_models/ggml-gpt4all-j-v1.3-groovy.bin
-MODEL_N_CTX=1000
-MODEL_N_BATCH=8
 
 #RESEND
 RESEND_API_KEY=
