+++ backend/llm/__init__.py
@@ -1,10 +1,12 @@
 from .base import BaseBrainPicking
+from .qa_base import QABaseBrainPicking
 from .openai import OpenAIBrainPicking
 from .openai_functions import OpenAIFunctionsBrainPicking
 from .private_gpt4all import PrivateGPT4AllBrainPicking
 
 __all__ = [
     "BaseBrainPicking",
+    "QABaseBrainPicking",
     "OpenAIBrainPicking",
     "OpenAIFunctionsBrainPicking",
     "PrivateGPT4AllBrainPicking",
